requirements.txt
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.24.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-vision==2.3.1
google-drive==0.3.2
google-drive-sync==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
Pillow==8.1.0
requests==2.25.1
Flask==1.1.2
opencv-python==4.5.2.52
oauth2client==4.1.3


jsonschema==3.2.0